whole lot of things split now works i hope minor issues with checking the score fixed (got a method for that now) also a bit of stuff with the dealer

+++ src/Karten.java
@@ -1,5 +1,5 @@
 public class Karten {
-    public int kartenwert;
+    public double kartenwert;
     public String suit;
     public String name;
